Adding Example for PlayStation 2

+++ Example/2_Basic_Primitive_Shape/GD101_Lib.h
@@ -21,11 +21,13 @@
 *
 **********************************************************************************************/
 
+
+
 #ifndef __GD_101_LIB__
 #define __GD_101_LIB__
 
-#include "GD101_DEF.h"
 
+#include "GD101_DEF.h"
 //--------- Core ---------------//////////////////////////////////////
 extern bool GD101_InitWindow(const char* windowName, int height, int width);
 extern int ScreenFlip();
@@ -36,9 +38,8 @@ extern bool EndLoop();
 extern int LoadShaderFile(const char * pFileName);
 extern void ExecuteShader();
 //extern int DrawLine(int x0, int y0, int x1, int y1);
-extern int DrawLine(int x0, int y0, int x1, int y1, COLOR4 color);
-extern int DrawBox(int x0, int y0, int width, int height, COLOR4 color);
-
+extern int DrawLine(float x0, float y0, float x1, float y1, COLOR4 color);
+extern int DrawBox(float x0, float y0, float width, float height, COLOR4 color);
 
 //------------- Inputs ----------/////////////////////////////////////
